+++ utils/logger_config.py
@@ -1,0 +1,8 @@
+import logging
+
+
+def setup_logging():
+    logging.basicConfig(
+        level=logging.DEBUG,
+        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+    )
